rename: {x}Props
