+++ packages/common/svelte/lib/svelteProps.ts
